+++ signac/__init__.py
@@ -31,9 +31,9 @@ from .core.h5store import H5Store, H5StoreManager
 from .core.jsondict import flush_all as flush
 from .db import get_database
 from .diff import diff_jobs
-from .synced_collections.buffered_collection import buffer_all as buffered
-from .synced_collections.buffered_collection import is_buffered
-from .synced_collections.collection_json import JSONDict
+from .synced_collections.backends.collection_json import JSONDict
+from .synced_collections.buffers.buffered_collection import buffer_all as buffered
+from .synced_collections.buffers.buffered_collection import is_buffered
 from .version import __version__
 
 __all__ = [
