+++ signac/synced_collections/buffers/__init__.py
@@ -1,0 +1,9 @@
+# Copyright (c) 2020 The Regents of the University of Michigan
+# All rights reserved.
+# This software is licensed under the BSD 3-Clause License.
+"""This subpackage defines the buffering protocol and supported implementations.
+
+No buffers are imported by default. Users should import desired buffers as
+needed.
+"""
+__all__ = [""]
