+++ tests/test_synced_collections/test_json_collection.py
@@ -8,7 +8,11 @@ from tempfile import TemporaryDirectory
 import pytest
 from synced_collection_test import SyncedDictTest, SyncedListTest
 
-from signac.synced_collections.collection_json import JSONCollection, JSONDict, JSONList
+from signac.synced_collections.backends.collection_json import (
+    JSONCollection,
+    JSONDict,
+    JSONList,
+)
 
 
 class JSONCollectionTest:
