+++ tests/test_synced_collections/test_redis_collection.py
@@ -7,7 +7,7 @@ import uuid
 import pytest
 from synced_collection_test import SyncedDictTest, SyncedListTest
 
-from signac.synced_collections.collection_redis import (
+from signac.synced_collections.backends.collection_redis import (
     RedisCollection,
     RedisDict,
     RedisList,
