+++ tests/test_synced_collections/test_utils.py
@@ -8,8 +8,8 @@ from tempfile import TemporaryDirectory
 
 import pytest
 
-from signac.synced_collections.collection_json import JSONDict
-from signac.synced_collections.synced_list import SyncedList
+from signac.synced_collections import SyncedList
+from signac.synced_collections.backends.collection_json import JSONDict
 from signac.synced_collections.utils import (
     AbstractTypeResolver,
     SyncedCollectionJSONEncoder,
