+++ tests/test_synced_collections/test_validators.py
@@ -3,7 +3,7 @@
 # This software is licensed under the BSD 3-Clause License.
 import pytest
 
-from signac.errors import InvalidKeyError, KeyTypeError
+from signac.synced_collections.errors import InvalidKeyError, KeyTypeError
 from signac.synced_collections.validators import json_format_validator, no_dot_in_key
 
 try:
