+++ tests/test_synced_collections/test_zarr_collection.py
@@ -6,7 +6,11 @@ from tempfile import TemporaryDirectory
 import pytest
 from synced_collection_test import SyncedDictTest, SyncedListTest
 
-from signac.synced_collections.collection_zarr import ZarrCollection, ZarrDict, ZarrList
+from signac.synced_collections.backends.collection_zarr import (
+    ZarrCollection,
+    ZarrDict,
+    ZarrList,
+)
 
 try:
     import numcodecs  # zarr depends on numcodecs
